update: copyEntryアクションをcopyEntriesアクションに変更.
