+++ src/features/storage/actions/copy-entries.ts
@@ -1,8 +1,12 @@
 "use server";
 
 import { getToken } from "@/actions/token";
-import { toCamelCase } from "@/lib/case-converters";
+import { toCamelCase, toSnakeCase } from "@/lib/case-converters";
 import { ActionError, ErrorResponse, toActionError } from "@/lib/errors";
+
+export type CopyEntryRequest = Readonly<{
+  key: string;
+}>;
 
 export type CopyEntryResponse = Readonly<{
   key: string;
@@ -12,9 +16,22 @@ export type CopyEntryResponse = Readonly<{
   updatedAt: Date;
 }>;
 
-export const copyEntry = async (
+export const copyEntries = async (
+  volumeName: string,
+  data: Record<string, CopyEntryRequest>,
+): Promise<
+  Record<string, { data?: CopyEntryResponse; error?: ActionError }>
+> => {
+  const res = await Promise.all(
+    Object.entries(data).map(([key, val]) => copyEntry(volumeName, key, val)),
+  );
+  return Object.fromEntries(Object.keys(data).map((key, i) => [key, res[i]]));
+};
+
+const copyEntry = async (
   volumeName: string,
   key: string,
+  data: CopyEntryRequest,
 ): Promise<{
   data?: CopyEntryResponse;
   error?: ActionError;
@@ -27,7 +44,9 @@ export const copyEntry = async (
         method: "POST",
         headers: {
           Authorization: `Session ${token}`,
+          "Content-Type": "application/json",
         },
+        body: JSON.stringify(toSnakeCase(data)),
         cache: "no-cache",
       },
     );
